+++ playhouse/apsw_ext.py
@@ -138,10 +138,6 @@ class APSWDatabase(SqliteDatabase):
         except AttributeError:
             return cursor.cursor.connection.changes()  # RETURNING query.
 
-    def begin(self, lock_type=None):
-        lock_type = lock_type or self._lock_type or 'deferred'
-        self.cursor().execute('begin %s;' % lock_type)
-
     def commit(self):
         if self.is_closed():
             raise InterfaceError('Cannot commit, database connection not '